+++ src/main/java/com/perfumeshop/dto/ProductRowDto.java
@@ -1,12 +1,14 @@
 package com.perfumeshop.dto;
 
+import java.math.BigDecimal;
+
 public class ProductRowDto {
     private Long id;
     private String name;
     private String category;
     private String brand;
     private Integer stock;
-    private Double price;
+    private BigDecimal price;
     private String image;
     private Boolean active;
     private String createdAt; // formatted string
@@ -28,8 +30,8 @@ public class ProductRowDto {
     public Integer getStock() { return stock; }
     public void setStock(Integer stock) { this.stock = stock; }
 
-    public Double getPrice() { return price; }
-    public void setPrice(Double price) { this.price = price; }
+    public BigDecimal getPrice() { return price; }
+    public void setPrice(BigDecimal price) { this.price = price; }
 
     public String getImage() { return image; }
     public void setImage(String image) { this.image = image; }
